+++ CHANGELOG.md
@@ -18,7 +18,8 @@
 
 ### :rocket: New Feature
 - `editor`
-  - [#1041](https://github.com/wix-incubator/rich-content/pull/1041) feat: shift+tab deletes tab character ('\t')
+  - [#1041](https://github.com/wix-incubator/rich-content/pull/1041) shift+tab click deletes tab character ('\t')
+  - [#1065](https://github.com/wix-incubator/rich-content/pull/1065) backspace click at start of block decrease indentation
 ### :bug: Bug Fix
 - `link-preview`
   - [#1048](https://github.com/wix-incubator/rich-content/pull/1048) fix: enable link preview and link embed by default